Descontructing homescreen
Componentization ...
