--- app/screens/HomeStyle.js
@@ -1,51 +1,0 @@
-const styles = {
-    iconAlign: { alignSelf: "center"},
-    drawerToggle: { padding: 20 },
-    headerStyle: { backgroundColor: "#fff"},
-    headerTitleStyle: {fontWeight: "bold"},
-    cardDetail: {
-        flex: 1,
-        padding: 10,
-        backgroundColor: "#fff",
-        justifyContent: "space-between"
-    },
-    containerDetail: {
-        flex: 1,
-        padding: 20,
-        borderTopWidth: 1,
-        marginRight: 10,
-        marginLeft: 10,
-        flexDirection: "column",
-        backgroundColor: "#fff",
-        justifyContent: "space-between"
-    },
-    cardSecSection:{
-        flex: 1,
-        padding: 10,
-        alignSelf: "center",
-        backgroundColor: "#fff",
-        justifyContent: "space-between"
-    },
-    lineInfo:{
-        justifyContent: "space-between",
-        flexDirection: "column"
-    },
-    lineSecSection:{
-        justifyContent: "space-between",
-        flexDirection: "row"
-    },
-    containerInfo: {
-        flex: 1,
-        flexDirection: "row",
-        justifyContent: "space-between"
-    },
-    confirmButton: {
-        flexDirection: "row",
-        justifyContent: "space-around",
-        backgroundColor: "#27ae60",
-        marginRight: 10,
-        marginLeft: 10
-    }
-}
-
-export default styles;
